next

+++ cgen.c
@@ -553,7 +553,7 @@ void processExprCall(DAST* dast, char* startLabel, char* endLabel) {
     
     DAST* args = dast->children[1];
 
-    for (int i = (int) (args->size - 1); i >= 0; i--) {
+    for (int i = args->size - 1; i >= 0; i--) {
         dispatch(args->children[i], startLabel, endLabel);
         emitADDI(SP, SP, -4);
         emitSW(S1, 0, SP);
